Move sitemarker into src/, small valgrind fixes.
Small tests with valgrind show open bugs
1. CSA is constructed but never desctructed
2. mismatched new/delete in seq_read
Will come back to these. Bigger issue:

On Pf chr10+mSP3.4 we see invalid reads when in bwt_seach_bwd.
Can't reproduce these on a smaller set. But anyway they
all seem to tie in with the list of pairs of lists of..
used for tracking which sites/alleles a read overlaps.
This is due to be replaced by sitemarker.